refactor(desktop): split markdown paper and optimize build output
