+++ apps/desktop/src/components/MarkdownPaper.tsx
@@ -1,58 +1,4 @@
-import { type CSSProperties, type Ref, type UIEvent, useCallback, useEffect, useMemo, useRef } from "react";
-import { defaultValueCtx, Editor, editorViewCtx, editorViewOptionsCtx, parserCtx, rootCtx, serializerCtx } from "@milkdown/kit/core";
-import { history } from "@milkdown/kit/plugin/history";
-import { listener, listenerCtx } from "@milkdown/kit/plugin/listener";
-import {
-  commands as commonmarkCommands,
-  imageSchema,
-  headingSchema,
-  inputRules as commonmarkInputRules,
-  keymap as commonmarkKeymap,
-  linkSchema,
-  paragraphSchema,
-  plugins as commonmarkPlugins,
-  schema as commonmarkSchema
-} from "@milkdown/kit/preset/commonmark";
-import {
-  commands as gfmCommands,
-  inputRules as gfmInputRules,
-  keymap as gfmKeymap,
-  pasteRules as gfmPasteRules,
-  plugins as gfmPlugins,
-  schema as gfmSchema
-} from "@milkdown/kit/preset/gfm";
-import { Milkdown, MilkdownProvider, useEditor, useInstance } from "@milkdown/react";
-import { Plugin } from "@milkdown/kit/prose/state";
-import type { EditorView } from "@milkdown/kit/prose/view";
-import { $prose } from "@milkdown/kit/utils";
-import { markraLiveMarkdownPlugin } from "@markra/editor";
-import {
-  markraClipboardImagePluginWithOptions,
-  type SaveClipboardImage,
-  type SaveRemoteClipboardImage
-} from "@markra/editor";
-import { markraCodeBlockPlugin } from "@markra/editor";
-import { markraCalloutPlugin } from "@markra/editor";
-import { markraCalloutSerializerPlugin } from "@markra/editor";
-import { markraHeadingSourcePlugin } from "@markra/editor";
-import { markraHeadingTogglePlugin } from "@markra/editor";
-import { normalizeHeadingSourceDocument } from "@markra/editor";
-import { markraLinkImageLivePlugin } from "@markra/editor";
-import { markraRawHtmlPlugin } from "@markra/editor";
-import { serializeLinkImageLiveMarkdown } from "@markra/editor";
-import { markraMarkdownShortcuts } from "@markra/editor";
-import { normalizeMarkdownShortcuts } from "@markra/editor";
-import { markraSlashCommands } from "@markra/editor";
-import { markraMathPlugin } from "@markra/editor";
-import { markraMathRemarkPlugin } from "@markra/editor";
-import { markraMathSourcePlugin } from "@markra/editor";
-import { markraTableControlsPlugin } from "@markra/editor";
-import { markraAiEditorPreviewPlugin } from "@markra/editor";
-import { markraAiSelectionHoldPlugin } from "@markra/editor";
-import { markraBlockDragPlugin } from "@markra/editor";
-import type { MarkdownShortcutMap } from "@markra/editor";
-import type { SlashCommandLabels } from "@markra/editor";
-import type { AiSelectionContext } from "@markra/ai";
+import { type CSSProperties, type Ref, type UIEvent } from "react";
 import { t, type AppLanguage } from "@markra/shared";
 import {
   editorContentWidthPixels,
@@ -61,31 +7,8 @@ import {
   type EditorContentWidth
 } from "../lib/editor-width";
 import type { EditorTheme } from "../lib/settings/app-settings";
-import { readAiSelectionContextFromView } from "../hooks/useEditorController";
-import type { MarkdownDocumentLinkFile } from "../lib/document-links";
-import { markraDocumentLinkCompletionPlugin } from "./document-link-completion";
 import { EditorWidthResizer } from "./EditorWidthResizer";
-
-const markraCommonmark = [
-  commonmarkSchema,
-  commonmarkInputRules,
-  commonmarkCommands,
-  commonmarkKeymap,
-  commonmarkPlugins
-].flat();
-
-const markraGfm = [
-  gfmSchema,
-  gfmInputRules,
-  gfmPasteRules,
-  gfmKeymap,
-  gfmCommands,
-  gfmPlugins
-].flat();
-
-function markdownShortcutSignature(shortcuts: MarkdownShortcutMap | undefined) {
-  return JSON.stringify(normalizeMarkdownShortcuts(shortcuts));
-}
+import { MarkdownPaperSurface, type MarkdownPaperSurfaceProps } from "./MarkdownPaperSurface";
 
 type MarkdownPaperProps = {
   autoFocus?: boolean;
@@ -95,428 +18,28 @@ type MarkdownPaperProps = {
   contentWidthMax?: number;
   contentWidthMin?: number;
   contentWidthPx?: number | null;
-  documentPath?: string | null;
+  documentPath?: MarkdownPaperSurfaceProps["documentPath"];
   editorTheme?: EditorTheme;
   initialContent: string;
   language?: AppLanguage;
   lineHeight?: number;
-  markdownShortcuts?: MarkdownShortcutMap;
-  onEditorReady: (editor: Editor | null, options?: { autoFocus?: boolean }) => unknown;
-  onMarkdownChange: (content: string) => unknown;
+  markdownShortcuts?: MarkdownPaperSurfaceProps["markdownShortcuts"];
+  onEditorReady: MarkdownPaperSurfaceProps["onEditorReady"];
+  onMarkdownChange: MarkdownPaperSurfaceProps["onMarkdownChange"];
   onContentWidthChange?: (width: number) => unknown;
   onContentWidthResizeEnd?: () => unknown;
   onContentWidthResizeStart?: () => unknown;
   onScroll?: (event: UIEvent<HTMLElement>) => unknown;
-  onSaveClipboardImage?: SaveClipboardImage;
-  onSaveRemoteClipboardImage?: SaveRemoteClipboardImage;
-  openExternalUrl?: (url: string) => unknown;
-  onTextSelectionChange?: (selection: AiSelectionContext | null) => unknown;
-  resolveImageSrc?: (src: string) => string;
+  onSaveClipboardImage?: MarkdownPaperSurfaceProps["onSaveClipboardImage"];
+  onSaveRemoteClipboardImage?: MarkdownPaperSurfaceProps["onSaveRemoteClipboardImage"];
+  openExternalUrl?: MarkdownPaperSurfaceProps["openExternalUrl"];
+  onTextSelectionChange?: MarkdownPaperSurfaceProps["onTextSelectionChange"];
+  resolveImageSrc?: MarkdownPaperSurfaceProps["resolveImageSrc"];
   revision: number;
   scrollRef?: Ref<HTMLElement>;
   topInset?: "tabs" | "titlebar";
-  workspaceFiles?: MarkdownDocumentLinkFile[];
+  workspaceFiles?: MarkdownPaperSurfaceProps["workspaceFiles"];
 };
-
-type MilkdownSurfaceProps = {
-  autoFocus: boolean;
-  documentPath?: MarkdownPaperProps["documentPath"];
-  initialContent: string;
-  language: AppLanguage;
-  onEditorReady: MarkdownPaperProps["onEditorReady"];
-  onMarkdownChange: (content: string) => unknown;
-  onSaveClipboardImage?: MarkdownPaperProps["onSaveClipboardImage"];
-  onSaveRemoteClipboardImage?: MarkdownPaperProps["onSaveRemoteClipboardImage"];
-  openExternalUrl?: MarkdownPaperProps["openExternalUrl"];
-  onTextSelectionChange?: MarkdownPaperProps["onTextSelectionChange"];
-  resolveImageSrc?: MarkdownPaperProps["resolveImageSrc"];
-  markdownShortcuts?: MarkdownPaperProps["markdownShortcuts"];
-  workspaceFiles?: MarkdownPaperProps["workspaceFiles"];
-};
-
-function markraTextSelectionObserverPlugin(
-  onTextSelectionChange: (selection: AiSelectionContext | null) => unknown
-) {
-  return $prose(() => {
-    let lastSignature = "";
-
-    const notifySelectionChange = (view: EditorView, options: { requireFocusForEmptySelection: boolean }) => {
-      const { selection } = view.state;
-
-      if (selection.empty) {
-        if (options.requireFocusForEmptySelection && !view.hasFocus()) return;
-
-        const blockContext = readAiSelectionContextFromView(view);
-        if (blockContext.text.trim()) {
-          const signature = `${blockContext.source ?? "block"}:${blockContext.from}:${blockContext.to}:${blockContext.text}`;
-          if (signature === lastSignature) return;
-
-          lastSignature = signature;
-          onTextSelectionChange(blockContext);
-          return;
-        }
-
-        if (lastSignature) {
-          lastSignature = "";
-          onTextSelectionChange(null);
-        }
-        return;
-      }
-
-      const text = view.state.doc.textBetween(selection.from, selection.to, "\n").trim();
-      if (!text) {
-        if (view.hasFocus() && lastSignature) {
-          lastSignature = "";
-          onTextSelectionChange(null);
-        }
-
-        return;
-      }
-
-      const signature = `${selection.from}:${selection.to}:${text}`;
-      if (signature === lastSignature) return;
-
-      lastSignature = signature;
-      onTextSelectionChange({
-        from: selection.from,
-        source: "selection",
-        text,
-        to: selection.to
-      });
-    };
-
-    const clearStaleSelectionAfterEditorClick = (view: EditorView) => {
-      if (!lastSignature) return;
-
-      const domSelection = view.dom.ownerDocument.getSelection();
-      const hasDomSelectedText = Boolean(domSelection && !domSelection.isCollapsed && domSelection.toString().trim());
-      if (hasDomSelectedText) return;
-
-      if (!view.state.selection.empty) {
-        lastSignature = "";
-        onTextSelectionChange(null);
-        return;
-      }
-
-      notifySelectionChange(view, { requireFocusForEmptySelection: false });
-    };
-
-    return new Plugin({
-      view(view) {
-        const ownerDocument = view.dom.ownerDocument;
-        const handleMouseUp = (event: MouseEvent) => {
-          if (event.button !== 0) return;
-
-          const targetElement =
-            event.target instanceof Element
-              ? event.target
-              : event.target instanceof Node
-                ? event.target.parentElement
-                : null;
-          const writingSurface = view.dom.closest(".paper-scroll");
-          if (!targetElement || !writingSurface?.contains(targetElement)) return;
-
-          ownerDocument.defaultView?.setTimeout(() => {
-            clearStaleSelectionAfterEditorClick(view);
-          }, 0);
-        };
-
-        ownerDocument.addEventListener("mouseup", handleMouseUp, true);
-
-        return {
-          destroy() {
-            ownerDocument.removeEventListener("mouseup", handleMouseUp, true);
-          },
-          update(view, previousState) {
-            const { selection } = view.state;
-            if (selection.eq(previousState.selection)) return;
-            notifySelectionChange(view, { requireFocusForEmptySelection: true });
-          }
-        };
-      }
-    });
-  });
-}
-
-function linkTargetFromClickTarget(target: EventTarget | null) {
-  const targetElement =
-    target instanceof Element ? target : target instanceof Node ? target.parentElement : null;
-  if (!targetElement) return null;
-
-  return targetElement.closest<HTMLAnchorElement | HTMLElement>("a[href], .markra-live-link-label[data-markra-href]");
-}
-
-function linkHrefFromClickTarget(target: EventTarget | null) {
-  const linkTarget = linkTargetFromClickTarget(target);
-  if (!linkTarget) return null;
-
-  if (linkTarget instanceof HTMLAnchorElement) {
-    return linkTarget.getAttribute("href") ?? linkTarget.href;
-  }
-
-  return linkTarget.dataset.markraHref ?? null;
-}
-
-function linkOpenModifierIsPressed(event: MouseEvent) {
-  return event.metaKey || event.ctrlKey;
-}
-
-function markraExternalLinkClickPlugin(openExternalUrl: (url: string) => unknown) {
-  return $prose(() => {
-    return new Plugin({
-      props: {
-        handleDOMEvents: {
-          mousedown(_view, event) {
-            const href = linkHrefFromClickTarget(event.target);
-            if (!href) return false;
-
-            if (!linkOpenModifierIsPressed(event)) {
-              return false;
-            }
-
-            event.preventDefault();
-            return true;
-          },
-          click(_view, event) {
-            const href = linkHrefFromClickTarget(event.target);
-            if (!href) return false;
-
-            if (!linkOpenModifierIsPressed(event)) {
-              return false;
-            }
-
-            event.preventDefault();
-
-            try {
-              Promise.resolve(openExternalUrl(href)).catch(() => {});
-            } catch {
-              // Opening external links is best-effort; editing should not be interrupted by opener failures.
-            }
-
-            return true;
-          }
-        }
-      }
-    });
-  });
-}
-
-function MilkdownInstanceBridge({ autoFocus, onEditorReady }: Pick<MilkdownSurfaceProps, "autoFocus" | "onEditorReady">) {
-  const [loading, getEditor] = useInstance();
-  const autoFocusRef = useRef(autoFocus);
-
-  useEffect(() => {
-    autoFocusRef.current = autoFocus;
-  }, [autoFocus]);
-
-  useEffect(() => {
-    if (loading) return;
-
-    const editor = getEditor();
-    onEditorReady(editor, { autoFocus: autoFocusRef.current });
-
-    return () => {
-      onEditorReady(null);
-    };
-  }, [getEditor, loading, onEditorReady]);
-
-  return null;
-}
-
-function MilkdownSurface({
-  autoFocus,
-  documentPath,
-  initialContent,
-  language,
-  onEditorReady,
-  onMarkdownChange,
-  onSaveClipboardImage,
-  onSaveRemoteClipboardImage,
-  openExternalUrl,
-  onTextSelectionChange,
-  resolveImageSrc,
-  markdownShortcuts,
-  workspaceFiles
-}: MilkdownSurfaceProps) {
-  const initialContentRef = useRef(initialContent);
-  const documentPathRef = useRef(documentPath);
-  const openExternalUrlRef = useRef(openExternalUrl);
-  const onSaveClipboardImageRef = useRef(onSaveClipboardImage);
-  const onSaveRemoteClipboardImageRef = useRef(onSaveRemoteClipboardImage);
-  const onTextSelectionChangeRef = useRef(onTextSelectionChange);
-  const workspaceFilesRef = useRef(workspaceFiles ?? []);
-  const externalLinkOpeningEnabled = Boolean(openExternalUrl);
-  const markdownDocumentLabel = t(language, "app.markdownDocument");
-  const tableControlLabels = {
-    addColumnRight: t(language, "editor.table.addColumnRight"),
-    addRowBelow: t(language, "editor.table.addRowBelow"),
-    alignLeft: t(language, "editor.table.alignLeft"),
-    alignCenter: t(language, "editor.table.alignCenter"),
-    alignRight: t(language, "editor.table.alignRight"),
-    deleteColumn: t(language, "editor.table.deleteColumn"),
-    deleteRow: t(language, "editor.table.deleteRow"),
-    adjustTable: t(language, "editor.table.adjustTable"),
-    resizeTableTo: t(language, "editor.table.resizeTableTo"),
-    tableColumns: t(language, "editor.table.columns"),
-    tableRows: t(language, "editor.table.rows")
-  };
-  const blockDragLabels = {
-    addBlock: t(language, "editor.blockAdd"),
-    dragBlock: t(language, "editor.blockDrag")
-  };
-  const headingToggleLabels = {
-    collapseSection: t(language, "editor.collapseSection"),
-    expandSection: t(language, "editor.expandSection")
-  };
-  const slashCommandLabels = useMemo<SlashCommandLabels>(() => ({
-    menu: t(language, "editor.slashCommands"),
-    noResults: t(language, "editor.slashCommandsNoResults"),
-    commands: {
-      bulletList: t(language, "menu.bulletList"),
-      callout: t(language, "menu.callout"),
-      codeBlock: t(language, "menu.codeBlock"),
-      heading1: t(language, "menu.heading1"),
-      heading2: t(language, "menu.heading2"),
-      heading3: t(language, "menu.heading3"),
-      orderedList: t(language, "menu.orderedList"),
-      paragraph: t(language, "menu.paragraph"),
-      quote: t(language, "menu.quote"),
-      table: t(language, "menu.table")
-    }
-  }), [language]);
-
-  useEffect(() => {
-    onSaveClipboardImageRef.current = onSaveClipboardImage;
-  }, [onSaveClipboardImage]);
-
-  useEffect(() => {
-    documentPathRef.current = documentPath;
-  }, [documentPath]);
-
-  useEffect(() => {
-    openExternalUrlRef.current = openExternalUrl;
-  }, [openExternalUrl]);
-
-  useEffect(() => {
-    onSaveRemoteClipboardImageRef.current = onSaveRemoteClipboardImage;
-  }, [onSaveRemoteClipboardImage]);
-
-  useEffect(() => {
-    onTextSelectionChangeRef.current = onTextSelectionChange;
-  }, [onTextSelectionChange]);
-
-  useEffect(() => {
-    workspaceFilesRef.current = workspaceFiles ?? [];
-  }, [workspaceFiles]);
-
-  const createEditor = useCallback(
-    (root: HTMLElement) => {
-      const editor = Editor.make()
-        .config((ctx) => {
-          ctx.set(rootCtx, root);
-          ctx.set(defaultValueCtx, initialContentRef.current);
-          ctx.update(editorViewOptionsCtx, (options) => ({
-            ...options,
-            attributes: {
-              ...options.attributes,
-              "aria-label": markdownDocumentLabel,
-              spellcheck: "true"
-            }
-          }));
-          ctx.get(listenerCtx).updated((editorCtx, doc) => {
-            try {
-              const view = editorCtx.get(editorViewCtx);
-              const normalizedDoc = normalizeHeadingSourceDocument(
-                view.state,
-                paragraphSchema.type(editorCtx),
-                headingSchema.type(editorCtx),
-                editorCtx.get(parserCtx)
-              );
-              onMarkdownChange(
-                serializeLinkImageLiveMarkdown(
-                  normalizedDoc === view.state.doc ? doc : normalizedDoc,
-                  editorCtx.get(serializerCtx),
-                  linkSchema.type(editorCtx),
-                  imageSchema.type(editorCtx)
-                )
-              );
-            } catch {
-              // Milkdown can flush a delayed update after teardown in tests or fast window closes.
-            }
-          });
-        })
-        .use(listener)
-        .use(history)
-        .use(markraMathRemarkPlugin)
-        .use(markraCommonmark)
-        .use(markraGfm)
-        .use(markraCalloutSerializerPlugin)
-        .use(markraCalloutPlugin)
-        .use(markraSlashCommands(slashCommandLabels))
-        .use(markraMathSourcePlugin)
-        .use(markraMarkdownShortcuts(markdownShortcuts))
-        .use(markraCodeBlockPlugin)
-        .use(markraMathPlugin)
-        .use(markraAiSelectionHoldPlugin)
-        .use(markraAiEditorPreviewPlugin)
-        .use(markraBlockDragPlugin(blockDragLabels))
-        .use(markraHeadingTogglePlugin(headingToggleLabels))
-        .use(
-          markraDocumentLinkCompletionPlugin({
-            getDocumentPath: () => documentPathRef.current,
-            getWorkspaceFiles: () => workspaceFilesRef.current
-          })
-        )
-        .use(
-          markraTextSelectionObserverPlugin((selection) => {
-            onTextSelectionChangeRef.current?.(selection);
-          })
-        )
-        .use(markraTableControlsPlugin(tableControlLabels))
-        .use(markraLinkImageLivePlugin(resolveImageSrc))
-        .use(markraHeadingSourcePlugin)
-        .use(
-          markraRawHtmlPlugin({
-            htmlSourceApplyLabel: t(language, "editor.htmlSourceApply"),
-            htmlSourceLabel: t(language, "editor.htmlSource"),
-            resolveImageSrc
-          })
-        )
-        .use(markraLiveMarkdownPlugin);
-
-      if (externalLinkOpeningEnabled) {
-        editor.use(
-          markraExternalLinkClickPlugin((url) => {
-            return openExternalUrlRef.current?.(url);
-          })
-        );
-      }
-
-      if (onSaveClipboardImageRef.current || onSaveRemoteClipboardImageRef.current) {
-        editor.use(
-          markraClipboardImagePluginWithOptions(
-            (image) => onSaveClipboardImageRef.current?.(image) ?? Promise.resolve(null),
-            {
-              saveRemoteImage: (image) => onSaveRemoteClipboardImageRef.current?.(image) ?? Promise.resolve(null)
-            }
-          )
-        );
-      }
-
-      return editor;
-    },
-    [externalLinkOpeningEnabled, language, markdownDocumentLabel, markdownShortcuts, onMarkdownChange, resolveImageSrc, slashCommandLabels]
-  );
-
-  useEditor(createEditor, [createEditor]);
-
-  return (
-    <>
-      <Milkdown />
-      <MilkdownInstanceBridge autoFocus={autoFocus} onEditorReady={onEditorReady} />
-    </>
-  );
-}
 
 export function MarkdownPaper({
   autoFocus = false,
@@ -555,11 +78,6 @@ export function MarkdownPaper({
     maxWidth: `${resolvedContentWidth}px`,
     ...(bottomOverlayInset > 0 ? { paddingBottom: `${bottomOverlayInset}px` } : {})
   } satisfies CSSProperties;
-  const shortcutsSignature = markdownShortcutSignature(markdownShortcuts);
-  const normalizedMarkdownShortcuts = useMemo(
-    () => normalizeMarkdownShortcuts(markdownShortcuts),
-    [shortcutsSignature]
-  );
   const topInsetClassName = topInset === "tabs" ? "pt-24 max-[900px]:pt-20" : "pt-14 max-[900px]:pt-10";
 
   return (
@@ -586,23 +104,21 @@ export function MarkdownPaper({
           onResizeEnd={onContentWidthResizeEnd}
           onResizeStart={onContentWidthResizeStart}
         />
-        <MilkdownProvider>
-          <MilkdownSurface
-            autoFocus={autoFocus}
-            documentPath={documentPath}
-            initialContent={initialContent}
-            language={language}
-            markdownShortcuts={normalizedMarkdownShortcuts}
-            onEditorReady={onEditorReady}
-            onMarkdownChange={onMarkdownChange}
-            onSaveClipboardImage={onSaveClipboardImage}
-            onSaveRemoteClipboardImage={onSaveRemoteClipboardImage}
-            openExternalUrl={openExternalUrl}
-            onTextSelectionChange={onTextSelectionChange}
-            resolveImageSrc={resolveImageSrc}
-            workspaceFiles={workspaceFiles}
-          />
-        </MilkdownProvider>
+        <MarkdownPaperSurface
+          autoFocus={autoFocus}
+          documentPath={documentPath}
+          initialContent={initialContent}
+          language={language}
+          markdownShortcuts={markdownShortcuts}
+          onEditorReady={onEditorReady}
+          onMarkdownChange={onMarkdownChange}
+          onSaveClipboardImage={onSaveClipboardImage}
+          onSaveRemoteClipboardImage={onSaveRemoteClipboardImage}
+          openExternalUrl={openExternalUrl}
+          onTextSelectionChange={onTextSelectionChange}
+          resolveImageSrc={resolveImageSrc}
+          workspaceFiles={workspaceFiles}
+        />
       </article>
     </section>
   );
