+++ apps/desktop/vite.config.ts
@@ -1,15 +1,19 @@
 import react from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
+import { fileURLToPath } from "node:url";
 import { defineConfig } from "vitest/config";
 import { stripDebugPlugin } from "./scripts/vite/strip-debug";
 
 const chunkSizeWarningLimitKb = 1200;
+const fontAssetPattern = /\.(?:otf|ttf|woff2?)$/i;
 const imageAssetPattern = /\.(?:avif|gif|ico|jpe?g|png|svg|webp)$/i;
+const browserNodeStubPath = fileURLToPath(new URL("./src/lib/browser-node-stub.ts", import.meta.url));
 
 function outputAssetFileName(asset: { names?: string[]; originalFileNames?: string[] }) {
   const sourceName = asset.names?.[0] ?? asset.originalFileNames?.[0] ?? "";
 
   if (imageAssetPattern.test(sourceName)) return "assets/images/[name]-[hash][extname]";
+  if (fontAssetPattern.test(sourceName)) return "assets/fonts/[name]-[hash][extname]";
 
   return "assets/[name]-[hash][extname]";
 }
@@ -39,6 +43,12 @@ const milkdownDependencies = dependencyPattern([
 const tauriDependencies = dependencyPattern(["@tauri-apps"]);
 const iconDependencies = dependencyPattern(["lucide-react", "lucide-static"]);
 const piAgentDependencies = dependencyPattern(["@mariozechner/pi-agent-core", "@mariozechner/pi-ai", "typebox"]);
+const codeEditorDependencies = dependencyPattern(["@codemirror", "codemirror"]);
+const dndDependencies = dependencyPattern(["@dnd-kit"]);
+const mathDependencies = dependencyPattern(["katex"]);
+const syntaxHighlightDependencies = dependencyPattern(["highlight.js", "lowlight"]);
+const toastDependencies = dependencyPattern(["sonner"]);
+const contentExtractionDependencies = dependencyPattern(["@mozilla/readability"]);
 const aiSdkDependencies = dependencyPattern([
   "@anthropic-ai",
   "@aws-sdk",
@@ -57,6 +67,12 @@ function vendorChunkName(id: string) {
   if (tauriDependencies.test(id)) return "tauri-vendor";
   if (iconDependencies.test(id)) return "icons-vendor";
   if (piAgentDependencies.test(id)) return "pi-agent-vendor";
+  if (codeEditorDependencies.test(id)) return "code-editor-vendor";
+  if (dndDependencies.test(id)) return "dnd-vendor";
+  if (mathDependencies.test(id)) return "math-vendor";
+  if (syntaxHighlightDependencies.test(id)) return "syntax-highlight-vendor";
+  if (toastDependencies.test(id)) return "toast-vendor";
+  if (contentExtractionDependencies.test(id)) return "content-extraction-vendor";
   if (aiSdkDependencies.test(id)) return "ai-sdk-vendor";
   if (id.includes("node_modules")) return "vendor";
 
@@ -66,6 +82,14 @@ function vendorChunkName(id: string) {
 export default defineConfig(({ mode }) => ({
   define: {
     __MARKRA_DEBUG__: JSON.stringify(mode !== "production")
+  },
+  resolve: {
+    alias: [
+      {
+        find: /^node:(?:fs|os|path)$/,
+        replacement: browserNodeStubPath
+      }
+    ]
   },
   plugins: [react(), tailwindcss(), ...(mode === "production" ? [stripDebugPlugin()] : [])],
   build: {
