+++ packages/ai/src/agent/tool-builders/anthropic.ts
@@ -1,6 +1,5 @@
 import type { Tool } from "@mariozechner/pi-ai";
-import type { AiProviderConfig } from "@markra/providers";
-import { getNativeWebSearchKind } from "@markra/providers";
+import { getNativeWebSearchKind, type AiProviderConfig } from "@markra/providers";
 
 export function buildAnthropicTools(
   config: AiProviderConfig,
