+++ packages/ai/src/agent/tool-builders/google.ts
@@ -1,5 +1,4 @@
-import type { AiProviderConfig } from "@markra/providers";
-import { getNativeWebSearchKind } from "@markra/providers";
+import { getNativeWebSearchKind, type AiProviderConfig } from "@markra/providers";
 
 export function buildGoogleTools(config: AiProviderConfig, model: string, webSearchEnabled: boolean | undefined) {
   return webSearchEnabled === true && getNativeWebSearchKind(config, model) === "google-search-grounding"
